ptn test refactor WIP
moving common code into test_support lib
