+++ test/test_ohClassPtn.cpp
@@ -1,5 +1,3 @@
-#include <chrono> //steady_clock, duration
-#include <thread> //this_thread
 #include <Omega_h_file.hpp>
 #include <Omega_h_library.hpp>
 #include <Omega_h_array_ops.hpp>
@@ -8,55 +6,9 @@
 #include <Omega_h_for.hpp>
 #include <redev_comm.h>
 #include "wdmcpl.h"
+#include "test_support.h"
 
-void timeMinMaxAvg(double time, double& min, double& max, double& avg) {
-  const auto comm = MPI_COMM_WORLD;
-  int nproc;
-  MPI_Comm_size(comm, &nproc);
-  double tot = 0;
-  MPI_Allreduce(&time, &min, 1, MPI_DOUBLE, MPI_MIN, comm);
-  MPI_Allreduce(&time, &max, 1, MPI_DOUBLE, MPI_MAX, comm);
-  MPI_Allreduce(&time, &tot, 1, MPI_DOUBLE, MPI_SUM, comm);
-  avg = tot / nproc;
-}
-
-void printTime(std::string mode, double min, double max, double avg) {
-  std::cout << mode << " elapsed time min, max, avg (s): "
-            << min << " " << max << " " << avg << "\n";
-}
-
-void getClassPtn(Omega_h::Mesh& mesh, redev::LOs& ranks, redev::LOs& classIds) {
-  auto ohComm = mesh.comm();
-  const auto dim = mesh.dim();
-  auto class_ids = mesh.get_array<Omega_h::ClassId>(dim, "class_id");
-  Omega_h::ClassId max_class = Omega_h::get_max(class_ids);
-  auto max_class_g = ohComm->allreduce(max_class,Omega_h_Op::OMEGA_H_MAX);
-  REDEV_ALWAYS_ASSERT(ohComm->size() == max_class_g);
-  auto class_ids_h = Omega_h::HostRead(class_ids);
-  if(!ohComm->rank()) {
-    //send ents with classId=2 to rank 1
-    Omega_h::Write<Omega_h::I32> ptnRanks(5,0);
-    Omega_h::Write<Omega_h::LO> ptnIdxs(5);
-    Omega_h::fill_linear(ptnIdxs,0,1);
-    auto owners = Omega_h::Remotes(ptnRanks, ptnIdxs);
-    mesh.migrate(owners);
-  } else {
-    int firstElm = 5;
-    const int elms = 18; // migrating elements [5:22]
-    Omega_h::Write<Omega_h::I32> ptnRanks(elms,0);
-    Omega_h::Write<Omega_h::LO> ptnIdxs(elms);
-    Omega_h::fill_linear(ptnIdxs,firstElm,1);
-    auto owners = Omega_h::Remotes(ptnRanks, ptnIdxs);
-    mesh.migrate(owners);
-  }
-
-  //the hardcoded assignment of classids to ranks
-  ranks.resize(3);
-  classIds.resize(3);
-  classIds[0] = 1; ranks[0] = 0;
-  classIds[1] = 2; ranks[1] = 1;
-  classIds[2] = 3; ranks[2] = 0;  //center ('O point') model vertex
-}
+namespace ts = test_support;
 
 void prepareMsg(Omega_h::Mesh& mesh, redev::ClassPtn& ptn,
     redev::LOs& dest, redev::LOs& offset, redev::LOs& permute) {
@@ -139,19 +91,6 @@ void getRdvPermutation(Omega_h::Mesh& mesh, redev::GO*& inGids, redev::GO inGids
   }
 }
 
-void checkAndAttachIds(Omega_h::Mesh& mesh, std::string name, redev::GO*& vtxData, redev::GOs& rdvPermute) {
-  auto gids_h = Omega_h::HostRead(mesh.globals(0));
-  Omega_h::HostWrite<Omega_h::GO> inVtxData_h(mesh.nverts());
-  for(int i=0; i<mesh.nverts(); i++)
-    inVtxData_h[i] = -1;
-  for(size_t i=0; i<rdvPermute.size(); i++) {
-    inVtxData_h[rdvPermute[i]] = vtxData[i];
-    REDEV_ALWAYS_ASSERT(gids_h[rdvPermute[i]] == vtxData[i]);
-  }
-  Omega_h::Write inVtxData(inVtxData_h);
-  mesh.add_tag(0,name,1,Omega_h::read(inVtxData));
-}
-
 int main(int argc, char** argv) {
   auto lib = Omega_h::Library(&argc, &argv);
   auto world = lib.world();
@@ -171,7 +110,7 @@ int main(int argc, char** argv) {
   if(isRdv) {
     //partition the omegah mesh by classification and return the
     //rank-to-classid array
-    getClassPtn(mesh, ranks, classIds);
+    ts::getClassPtn(mesh, ranks, classIds);
     REDEV_ALWAYS_ASSERT(ranks.size()==3);
     REDEV_ALWAYS_ASSERT(ranks.size()==classIds.size());
     Omega_h::vtk::write_parallel("rdvSplit.vtk", &mesh, mesh.dim());
@@ -214,13 +153,7 @@ int main(int argc, char** argv) {
       auto start = std::chrono::steady_clock::now();
       comm.Pack(appOutDest, appOutOffsets, msgs.data());
       comm.Send();
-      auto end = std::chrono::steady_clock::now();
-      std::chrono::duration<double> elapsed_seconds = end-start;
-      double min, max, avg;
-      timeMinMaxAvg(elapsed_seconds.count(), min, max, avg);
-      if( iter == 0 ) ss << "write";
-      std::string str = ss.str();
-      if(!rank) printTime(str, min, max, avg);
+      ts::getAndPrintTime(start,name + " write",rank);
     } else {
       redev::GO* msgs;
       auto start = std::chrono::steady_clock::now();
@@ -228,25 +161,18 @@ int main(int argc, char** argv) {
       comm.Unpack(rdvSrcRanks,rdvOffsets,msgs,msgStart,msgCount,knownSizes);
       REDEV_ALWAYS_ASSERT(rdvOffsets == redev::GOs({0,6,19}));
       REDEV_ALWAYS_ASSERT(rdvSrcRanks == redev::GOs({0,0}));
+      redev::GOs msgVec(msgs, msgs+msgCount);
       if(!rank) {
         REDEV_ALWAYS_ASSERT(msgStart==0 && msgCount==6);
-        redev::GOs msgVec(msgs, msgs+msgCount);
         REDEV_ALWAYS_ASSERT(msgVec == redev::GOs({0,2,3,4,5,6}));
       } else {
         REDEV_ALWAYS_ASSERT(msgStart==6 && msgCount==13);
-        redev::GOs msgVec(msgs, msgs+msgCount);
         REDEV_ALWAYS_ASSERT(msgVec == redev::GOs({1,7,8,9,10,11,12,13,14,15,16,17,18}));
       }
-      auto end = std::chrono::steady_clock::now();
-      std::chrono::duration<double> elapsed_seconds = end-start;
-      double min, max, avg;
-      timeMinMaxAvg(elapsed_seconds.count(), min, max, avg);
-      if( iter == 0 ) ss << "read";
-      std::string str = ss.str();
-      if(!rank) printTime(str, min, max, avg);
+      ts::getAndPrintTime(start,name + " read",rank);
       //attach the ids to the mesh
       if(iter==0) getRdvPermutation(mesh, msgs, msgCount, rdvInPermute);
-      checkAndAttachIds(mesh, "inVtxGids", msgs, rdvInPermute);
+      ts::checkAndAttachIds(mesh, "inVtxGids", msgVec, rdvInPermute);
       Omega_h::vtk::write_parallel("rdvInGids.vtk", &mesh, mesh.dim());
       delete [] msgs;
     } //end non-rdv -> rdv
