+++ test/test_support.cpp
@@ -1,0 +1,70 @@
+#include "test_support.h"
+#include <mpi.h>
+
+namespace test_support {
+
+void timeMinMaxAvg(double time, double& min, double& max, double& avg) {
+  const auto comm = MPI_COMM_WORLD;
+  int nproc;
+  MPI_Comm_size(comm, &nproc);
+  double tot = 0;
+  MPI_Allreduce(&time, &min, 1, MPI_DOUBLE, MPI_MIN, comm);
+  MPI_Allreduce(&time, &max, 1, MPI_DOUBLE, MPI_MAX, comm);
+  MPI_Allreduce(&time, &tot, 1, MPI_DOUBLE, MPI_SUM, comm);
+  avg = tot / nproc;
+}
+
+void printTime(std::string mode, double min, double max, double avg) {
+  std::cout << mode << " elapsed time min, max, avg (s): "
+            << min << " " << max << " " << avg << "\n";
+}
+
+void getClassPtn(Omega_h::Mesh& mesh, redev::LOs& ranks, redev::LOs& classIds) {
+  auto ohComm = mesh.comm();
+  const auto dim = mesh.dim();
+  auto class_ids = mesh.get_array<Omega_h::ClassId>(dim, "class_id");
+  Omega_h::ClassId max_class = Omega_h::get_max(class_ids);
+  auto max_class_g = ohComm->allreduce(max_class,Omega_h_Op::OMEGA_H_MAX);
+  REDEV_ALWAYS_ASSERT(ohComm->size() == max_class_g);
+  auto class_ids_h = Omega_h::HostRead(class_ids);
+  if(!ohComm->rank()) {
+    //send ents with classId=2 to rank 1
+    Omega_h::Write<Omega_h::I32> ptnRanks(5,0);
+    Omega_h::Write<Omega_h::LO> ptnIdxs(5);
+    Omega_h::fill_linear(ptnIdxs,0,1);
+    auto owners = Omega_h::Remotes(ptnRanks, ptnIdxs);
+    mesh.migrate(owners);
+  } else {
+    int firstElm = 5;
+    const int elms = 18; // migrating elements [5:22]
+    Omega_h::Write<Omega_h::I32> ptnRanks(elms,0);
+    Omega_h::Write<Omega_h::LO> ptnIdxs(elms);
+    Omega_h::fill_linear(ptnIdxs,firstElm,1);
+    auto owners = Omega_h::Remotes(ptnRanks, ptnIdxs);
+    mesh.migrate(owners);
+  }
+
+  //the hardcoded assignment of classids to ranks
+  ranks.resize(3);
+  classIds.resize(3);
+  classIds[0] = 1; ranks[0] = 0;
+  classIds[1] = 2; ranks[1] = 1;
+  classIds[2] = 3; ranks[2] = 0;  //center ('O point') model vertex
+}
+
+void checkAndAttachIds(Omega_h::Mesh& mesh, std::string name, redev::GOs& vtxData, redev::GOs& rdvPermute) {
+  REDEV_ALWAYS_ASSERT(rdvPermute.size() == vtxData.size());
+  auto gids_h = Omega_h::HostRead(mesh.globals(0));
+  Omega_h::HostWrite<Omega_h::GO> inVtxData_h(mesh.nverts());
+  for(int i=0; i<mesh.nverts(); i++)
+    inVtxData_h[i] = -1;
+  for(size_t i=0; i<rdvPermute.size(); i++) {
+    inVtxData_h[rdvPermute[i]] = vtxData[i];
+    REDEV_ALWAYS_ASSERT(gids_h[rdvPermute[i]] == vtxData[i]);
+  }
+  Omega_h::Write inVtxData(inVtxData_h);
+  mesh.add_tag(0,name,1,Omega_h::read(inVtxData));
+  mesh.sync_tag(0,name);
+}
+
+}
